pyproject.toml
[tool.poetry]
name = "NF caravans"
version = "1.0"
description = "NF caravans code, made by the [Giant Digital](https://www.giantdigital.co.uk) team."
authors = [
    "Jon Atkinson <jon.atkinson@giantdigital.co.uk>",
    "Peter Smith-Keary <peter.smith-keary@giantdigital.co.uk>",
    "Wayne Fagin <wayne.fagin@giantdigital.co.uk>",
    "Tom Bannister <tom.bannister@giantdigital.co.uk>",
    "Will Hoey <will.hoey@giantdigital.co.uk>",
    "Dominic Chaple <dominic.chaple@giantdigital.co.uk>",
    "Ricardo Mendonca <ricardo.mendonca@giantdigital.co.uk>",
    "Scott Pashley <scott.pashley@giantdigital.co.uk>",
    "Jake Quirke <jake.quirke@giantdigital.co.uk>",
    "Karl Swatman <karl.swatman@giantdigital.co.uk>",
    "Toby Lees <toby.lees@giantdigital.co.uk>",
    "Rob Tagg <rob.tagg@giantdigital.co.uk"
]

[tool.poetry.dependencies]
python = "~3.11"
django = "~4"
django-cms = "~3"
django-sekizai = "~2"
pillow = "^9.3"
sentry-sdk = "^1.14"
dj-database-url = "^0.5.0"
django-cache-url = "^3.2"
gunicorn = "^20.0.4"
django-redis = "^5.0"
dj-email-url = "^1.0.2"
python-environ = "^0.4.54"
packaging = "^23.0"
giant-mixins = "^0.3.5"
giant-news = "^1.2.3"
form-designer = "^0.21.2"
django-admin-extra-buttons = "^1.5.6"
django-admin-ordering = "^0.17.0"
faker = "^18.10.1"
django-taggit = "^4.0.0"
django-taggit-autosuggest = "^0.4.1"
giant-search = "^1.0.0"
django-import-export = "^4.2.0"
django-q2 = "^1.7.3"

[tool.poetry.group.dev.dependencies]
black = "^23.1"
coverage = "^7.2"
flake8 = "^6.0"
pytest = "^7.2"
pytest-cov = "^4.0"
pytest-django = "^4.5"
pytest-mock = "^3.10"
flake8-bugbear = "^23.3"
flake8-builtins = "^2.1"
flake8-debugger = "^4.1"
flake8-print = "^5.0"
flake8-quotes = "^3.3"
flake8-isort = "^6.0"
isort = "^5.1"
factory-boy = "^3.2.1"
ipython = "8.27.0"

[tool.black]
line-length = 99
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.ipython
  | \.mypy_cache
  | \.ropeproject
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
